+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # nightly
+- Add `Simple Remove Think` node
 # v1.0.6
 - Add `Simple Auto Bypass` node
 - Fix "always run" mode for `Simple Lora Loader`